## About The Pull Request
Adds an Anteater Mutation Toxin to the Troutstation files specifically, as well as a Green Slime Extract reaction to create this toxin. This follows the implementation of the Lizardperson Mutation Toxin.

For this reaction, the reagent to inject into the green slime extract is ants.

## Why It's Good For The Game

Consistency!
* All the other roundstart races have mutation toxins.
* Some of the other roundstart races have Green Slime Extract ways to produce the mutation toxins.
** Note: Upstream, people apparently complain a ton if they're turned into felinids. I doubt that's a problem for this environment.
** There's currently a PR to add felinids and moth folks to the Green Slime Extract list. If that gets closed, I can pretty easily add it to this station.

## Changelog

:cl: Kaltren
add: Anteater Mutation Toxin now exists as a reagent. It cannot show up in random recipes.
add: Injecting ants into a Green Slime Extract now produces Anteater Mutation Toxin.
/:cl: